+++ backend/src/main/java/com/kop/morning/domain/article/article/dto/requestDto/ArticleRequestDto.java
@@ -1,4 +1,4 @@
-package com.kop.morning.domain.article.article.dto.responseDto;
+package com.kop.morning.domain.article.article.dto.requestDto;
 
 import lombok.Data;
