+++ backend/src/main/java/com/kop/morning/domain/article/article/entity/Article.java
@@ -1,7 +1,7 @@
 package com.kop.morning.domain.article.article.entity;
 
 import com.kop.morning.domain.article.article.dto.requestDto.ArticleUpdateRequestDto;
-import com.kop.morning.domain.article.article.dto.responseDto.ArticleRequestDto;
+import com.kop.morning.domain.article.article.dto.requestDto.ArticleRequestDto;
 import com.kop.morning.domain.article.comment.entity.Comment;
 import com.kop.morning.domain.member.entity.Member;
 import jakarta.persistence.CascadeType;
